# openapi-spec.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/v1/assignments/{id}": {
            "get": {
                "tags": [
                    "assignment-controller"
                ],
                "operationId": "getAssignmentById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Assignment"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "assignment-controller"
                ],
                "operationId": "updateAssignment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Assignment"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Assignment"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "assignment-controller"
                ],
                "operationId": "deleteAssignment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "boolean"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/wages/wage": {
            "post": {
                "tags": [
                    "wage-controller"
                ],
                "operationId": "createAssignment",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Wage"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/assignments": {
            "get": {
                "tags": [
                    "assignment-controller"
                ],
                "operationId": "getAllAssignments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Assignment"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "assignment-controller"
                ],
                "operationId": "createAssignment_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Assignment"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Assignment"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/bills/getBillsForProject": {
            "get": {
                "tags": [
                    "bills-controller"
                ],
                "operationId": "getBillsForProject",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Bills"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/bills/getBillsForMonthAndYear": {
            "get": {
                "tags": [
                    "bills-controller"
                ],
                "operationId": "getBillsForMonthAndYear",
                "parameters": [
                    {
                        "name": "selectedDate",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Bills"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/bills/getBillsForEmployee": {
            "get": {
                "tags": [
                    "bills-controller"
                ],
                "operationId": "getBillsForEmployee",
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Bills"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/assignmentsForProject": {
            "get": {
                "tags": [
                    "assignment-controller"
                ],
                "operationId": "getAssignmentByProjectId",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Assignment"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Assignment": {
                "type": "object",
                "properties": {
                    "assignmentId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "projectId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "assignmentType": {
                        "type": "string"
                    },
                    "assignmentTaxType": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "wage": {
                        "type": "number",
                        "format": "float"
                    },
                    "status": {
                        "type": "string"
                    },
                    "employeeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "description": {
                        "type": "string"
                    },
                    "assignmentNotes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Notes"
                        }
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
            "Customer": {
                "type": "object",
                "properties": {
                    "customerId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "customerCompanyName": {
                        "type": "string"
                    },
                    "customerName": {
                        "type": "string"
                    },
                    "customerPhone": {
                        "type": "string"
                    },
                    "customerEmail": {
                        "type": "string"
                    },
                    "ein": {
                        "type": "string"
                    },
                    "website": {
                        "type": "string"
                    },
                    "customerStatus": {
                        "type": "string"
                    },
                    "customerStartDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "customerEndDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "customerContactEmail": {
                        "type": "string"
                    },
                    "customerType": {
                        "type": "string"
                    },
                    "customerAddress": {
                        "type": "string"
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
...
            "Notes": {
                "type": "object",
                "properties": {
                    "noteId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "notesType": {
                        "type": "string"
                    },
                    "details": {
                        "type": "string"
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
            "Project": {
                "type": "object",
                "properties": {
                    "projectId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "projectName": {
                        "type": "string"
                    },
                    "customer": {
                        "$ref": "#/components/schemas/Customer"
                    },
                    "employee": {
                        "$ref": "#/components/schemas/Employee"
                    },
                    "client": {
                        "type": "string"
                    },
                    "billRates": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Wage"
                        }
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "invoiceTerm": {
                        "type": "string"
                    },
                    "paymentTerm": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "assignments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Assignment"
                        }
                    },
                    "employeeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "vendorId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
...
            "Wage": {
                "type": "object",
                "properties": {
                    "wageId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "wage": {
                        "type": "number",
                        "format": "float"
                    },
                    "wageType": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "wageNotes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Notes"
                        }
                    },
                    "createdDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
...
            "Bills": {
                "type": "object",
                "properties": {
                    "invoiceId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "assignmentId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "invoiceMonth": {
                        "type": "string",
                        "format": "date"
                    },
                    "billing": {
                        "type": "number",
                        "format": "float"
                    },
                    "employeeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "employeeName": {
                        "type": "string"
                    },
                    "hours": {
                        "type": "number",
                        "format": "float"
                    },
                    "total": {
                        "type": "number",
                        "format": "float"
                    },
                    "billPaidAmount": {
                        "type": "number",
                        "format": "float"
                    },
                    "billDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "paymentDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "status": {
                        "type": "string"
                    },
                    "billType": {
                        "type": "string"
                    },
                    "billId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            }
        }
    }
}